Move ex_trust to executor.c
