+++ src/nvim/lua/executor.c
@@ -20,6 +20,7 @@
 #include "nvim/buffer_defs.h"
 #include "nvim/change.h"
 #include "nvim/cursor.h"
+#include "nvim/charset.h"
 #include "nvim/drawscreen.h"
 #include "nvim/eval.h"
 #include "nvim/eval/funcs.h"
@@ -2218,30 +2219,45 @@ char *nlua_read_secure(const char *path)
   return buf;
 }
 
-char *nlua_trust(const char *path, const char *mode)
+void ex_trust(exarg_T *eap)
 {
-  lua_State *const lstate = global_lstate;
-  lua_getglobal(lstate, "vim");
-  lua_getfield(lstate, -1, "secure");
-  lua_getfield(lstate, -1, "trust");
-  lua_pushstring(lstate, path);
-  lua_pushstring(lstate, mode);
-  if (nlua_pcall(lstate, 2, 2)) {
-    nlua_error(lstate, _("Error executing vim.secure.trust: %.*s"));
-    return NULL;
+  char *mode;
+  char *path;
+  {
+    char *p = skiptowhite(eap->arg);
+    size_t len = (size_t)(p - eap->arg);
+    mode = xcalloc(len + 1, sizeof(char));
+    memcpy(mode, eap->arg, len);
+    path = skipwhite(p);
   }
 
-  char *errmsg = NULL;
-  size_t len = 0;
-  const char *contents = lua_tolstring(lstate, -1, &len);
-  if (contents != NULL) {
-    // Add one to include trailing null byte
-    errmsg = xcalloc(len + 1, sizeof(char));
-    memcpy(errmsg, contents, len + 1);
+  {
+    lua_State *const lstate = global_lstate;
+    lua_getglobal(lstate, "vim");
+    lua_getfield(lstate, -1, "secure");
+    lua_getfield(lstate, -1, "trust");
+    lua_pushstring(lstate, path);
+    lua_pushstring(lstate, mode);
+    if (nlua_pcall(lstate, 2, 2)) {
+      nlua_error(lstate, _("Error executing vim.secure.trust: %.*s"));
+      return;
+    }
+
+    size_t len = 0;
+    const char *contents = lua_tolstring(lstate, -1, &len);
+    if (contents != NULL) {
+      // Add one to include trailing null byte
+      char *errmsg = xcalloc(len + 1, sizeof(char));
+      memcpy(errmsg, contents, len + 1);
+      if (errmsg != NULL) {
+        semsg(errmsg);
+        xfree(errmsg);
+      }
+    }
+
+    // Pop return values, "vim" and "secure"
+    lua_pop(lstate, 4);
   }
 
-  // Pop return values, "vim" and "secure"
-  lua_pop(lstate, 4);
-
-  return errmsg;
+  xfree(mode);
 }
